requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
python-dotenv
openai
PyMuPDF